AArch64: Add compensation code to TR_UnloadedClassPicSite
Add compensation code to TR_UnloadedClassPicSite for aarch64.
This code uses _size member to check if the site is for address
materialization sequence.
Unlike arm or ppc, the lsb of the mov instruction of aarch64
is used for encoding the target register number. Thus, we cannot simply
set lsb of the patch location to invalidate it.
If _size is 4, we patch the pic location to `movz rX, #1`. Otherwise,
we store -1 to the pic location.

Signed-off-by: Akira Saitoh <saiaki@jp.ibm.com>